requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pydub>=0.25.1
requests>=2.31.0
httpx>=0.24.0

# Tencent Cloud SDK for TTS and Hunyuan
tencentcloud-sdk-python>=3.0.900

# Search and content extraction
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
readability-lxml>=0.8.1
playwright>=1.40.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Configuration
configparser>=5.3.0
python-dotenv>=1.0.0

# Logging
colorlog>=6.7.0

# Tencent Cloud COS for audio storage
cos-python-sdk-v5>=1.9.0
